Editor: Move the Excerpt panel to the editor package
