+++ packages/edit-post/src/components/sidebar/settings-sidebar/index.js
@@ -13,6 +13,7 @@ import { store as interfaceStore } from '@wordpress/interface';
 import { store as keyboardShortcutsStore } from '@wordpress/keyboard-shortcuts';
 import {
 	store as editorStore,
+	PostExcerptPanel,
 	PostFeaturedImagePanel,
 	PostLastRevisionPanel,
 	PostTaxonomiesPanel,
@@ -23,7 +24,6 @@ import {
  */
 import SettingsHeader from '../settings-header';
 import PostStatus from '../post-status';
-import PostExcerpt from '../post-excerpt';
 import DiscussionPanel from '../discussion-panel';
 import PageAttributes from '../page-attributes';
 import MetaBoxes from '../../meta-boxes';
@@ -84,7 +84,7 @@ const SidebarContent = ( {
 							<PostLastRevisionPanel />
 							<PostTaxonomiesPanel />
 							<PostFeaturedImagePanel />
-							<PostExcerpt />
+							<PostExcerptPanel />
 							<DiscussionPanel />
 							<PageAttributes />
 							<MetaBoxes location="side" />
