+++ packages/edit-site/src/components/sidebar-edit-mode/page-panels/index.js
@@ -13,6 +13,7 @@ import { useSelect } from '@wordpress/data';
 import { store as coreStore } from '@wordpress/core-data';
 import { decodeEntities } from '@wordpress/html-entities';
 import {
+	PostExcerptPanel,
 	PostFeaturedImagePanel,
 	PostLastRevisionPanel,
 	PostTaxonomiesPanel,
@@ -102,6 +103,7 @@ export default function PagePanels() {
 			<PostLastRevisionPanel />
 			<PostTaxonomiesPanel />
 			<PostFeaturedImagePanel />
+			<PostExcerptPanel />
 		</>
 	);
 }
