+++ packages/edit-site/src/components/sidebar-edit-mode/template-panel/index.js
@@ -4,6 +4,7 @@
 import { useSelect } from '@wordpress/data';
 import { PanelBody } from '@wordpress/components';
 import {
+	PostExcerptPanel,
 	PostFeaturedImagePanel,
 	PostLastRevisionPanel,
 	PostTaxonomiesPanel,
@@ -66,6 +67,7 @@ export default function TemplatePanel() {
 			<PostLastRevisionPanel />
 			<PostTaxonomiesPanel />
 			<PostFeaturedImagePanel />
+			<PostExcerptPanel />
 		</>
 	);
 }
