+++ packages/editor/src/components/post-excerpt/plugin.js
