+++ packages/editor/src/private-apis.js
@@ -8,6 +8,7 @@ import { EntitiesSavedStatesExtensible } from './components/entities-saved-state
 import useBlockEditorSettings from './components/provider/use-block-editor-settings';
 import PostPanelRow from './components/post-panel-row';
 import PreviewDropdown from './components/preview-dropdown';
+import PluginPostExcerpt from './components/post-excerpt/plugin';
 
 export const privateApis = {};
 lock( privateApis, {
@@ -16,6 +17,7 @@ lock( privateApis, {
 	EntitiesSavedStatesExtensible,
 	PostPanelRow,
 	PreviewDropdown,
+	PluginPostExcerpt,
 
 	// This is a temporary private API while we're updating the site editor to use EditorProvider.
 	useBlockEditorSettings,
